To test parameters given to alterImage()-function
Change in X or Y coordinates while dragging divided by time change in milliseconds gives value in range of 0 to 3.5 or 4 (abs values). To reach those large values mouse needs to be dragged really fast over the edge of processing window. So it's reasonable to limit max values to 1 to give user a good control on gradient length.